requirements.txt
imagingcontrol4>=1.2.0
imagingcontrol4pyside6
opencv-python
numpy
pymmcore==10.1.1.70.6